refactor: 검색결과 미리보기 컴포넌트 구조 단순화
