+++ src/asset/css/BookSearchPreview.css
@@ -6,17 +6,18 @@
 
 .search-preview__list {
   display: flex;
+  overflow: hidden;
+  width: 100%;
+}
+
+.search-preview__books {
+  display: flex;
   flex-direction: column;
   flex-grow: 1;
   flex-shrink: 1;
   overflow: hidden;
 }
 
-.search-preview__left {
-  display: flex;
-  overflow: hidden;
-  width: 100%;
-}
 
 .search-preview__book {
   text-align: left;
