+++ src/component/utils/BookSearchPreview.tsx
@@ -1,7 +1,10 @@
-import { useGetSearchKeywordsAutocomplete } from "~/api/searchKeywords/useGetSearchKeywordsAutoComplete";
-import { BookInfo } from "~/type";
+import { useState } from "react";
+import { Link } from "react-router-dom";
+import { useGetSearchKeywordsAutocomplete } from "~/api/searchKeywords/useGetSearchKeywordsAutocomplete";
+import { type BookInfo } from "~/type";
 import BookSearchPreviewList from "~/component/utils/BookSearchPreviewList";
-import EmphasisInString from "./EmphasisInString";
+import PaginationCircle from "~/component/utils/PaginationCircle";
+import "~/asset/css/BookSearchPreview.css";
 
 export type BookPreviewType = Omit<BookInfo, "category">;
 
@@ -9,36 +12,30 @@ type Props = {
   keyword: string;
 };
 
+const PAGE_SIZE = 3;
+
 const BookSearchPreview = ({ keyword }: Props) => {
   const { books } = useGetSearchKeywordsAutocomplete();
+  const [page, setPage] = useState(0);
+
+  const slicedBooks = books.slice(page * PAGE_SIZE, (page + 1) * PAGE_SIZE);
+  const pageCount = Math.floor(books.length / PAGE_SIZE);
 
   return (
     <div className="search-preview__wrapper">
-      <BookSearchPreviewList
-        keyword={keyword}
-        books={books}
-        bookUI={({ book }) => {
-          return (
-            <>
-              <p className="search-preview__book__title">
-                <EmphasisInString wholeString={book.title} emphasis={keyword} />
-              </p>
-              <span className="search-preview__book__author">
-                <EmphasisInString
-                  wholeString={book.author}
-                  emphasis={keyword}
-                />
-              </span>
-              <span className="search-preview__book__publisher">
-                <EmphasisInString
-                  wholeString={book.publisher}
-                  emphasis={keyword}
-                />
-              </span>
-            </>
-          );
-        }}
+      <BookSearchPreviewList keyword={keyword} books={slicedBooks} />
+      <PaginationCircle
+        className="search-preview__pagination"
+        currentPage={page}
+        setCurrentPage={setPage}
+        length={pageCount}
       />
+      <Link
+        className="search-preview__more"
+        to={`/search?search=${encodeURI(keyword)}`}
+      >
+        {`검색 결과 더보기`}
+      </Link>
     </div>
   );
 };
