+++ src/component/utils/BookSearchPreviewList.tsx
@@ -1,74 +1,58 @@
-import { MouseEvent, useEffect, useState } from "react";
-import { useNavigate, Link } from "react-router-dom";
+import { useState } from "react";
+import { useNavigate } from "react-router-dom";
 import { BookPreviewType } from "~/component/utils/BookSearchPreview";
+import EmphasisInString from "~/component/utils/EmphasisInString";
 import Image from "~/component/utils/Image";
-import PaginationCircle from "~/component/utils/PaginationCircle";
-import "~/asset/css/BookSearchPreview.css";
 
 type Props = {
   keyword: string;
   books: BookPreviewType[];
-  bookUI: (props: { book: BookPreviewType }) => JSX.Element;
 };
 
-const PAGE_SIZE = 3;
-const BookSearchPreviewList = ({ keyword, books, bookUI }: Props) => {
-  const [page, setPage] = useState(0);
+const BookSearchPreviewList = ({ keyword, books }: Props) => {
   const [selectedIndex, setSelectedIndex] = useState(0);
-
-  const slicedBooks = books.slice(page * PAGE_SIZE, (page + 1) * PAGE_SIZE);
-  const pageCount = Math.floor(books.length / PAGE_SIZE);
   const navigate = useNavigate();
 
   return (
-    <>
-      <div className="search-preview__left">
-        <div className="search-preview__list">
-          {slicedBooks.map((book, index) => {
-            const isSelected = index === selectedIndex;
-
-            return (
-              <button
-                type="button"
-                key={book.id}
-                value={book.id}
-                className={`search-preview__book ${
-                  isSelected ? "selected" : ""
-                }`}
-                onClick={e => {
-                  e.preventDefault();
-                  navigate(`/info/${book.id}`);
-                }}
-                onMouseOver={() => setSelectedIndex(index)}
-              >
-                {bookUI({ book })}
-              </button>
-            );
-          })}
-        </div>
-        <Image
-        className="search-preview__image"
-          width={120}
-          height={180}
-          src={slicedBooks[selectedIndex]?.image}
-          alt={slicedBooks[selectedIndex]?.title}
-        />
+    <div className="search-preview__list">
+      <div className="search-preview__books">
+        {books.map((book, index) => (
+          <button
+            type="button"
+            key={book.id}
+            value={book.id}
+            className={`search-preview__book ${
+              index === selectedIndex ? "selected" : ""
+            }`}
+            onClick={e => {
+              e.preventDefault();
+              navigate(`/info/${book.id}`);
+            }}
+            onMouseOver={() => setSelectedIndex(index)}
+          >
+            <p className="search-preview__book__title">
+              <EmphasisInString wholeString={book.title} emphasis={keyword} />
+            </p>
+            <span className="search-preview__book__author">
+              <EmphasisInString wholeString={book.author} emphasis={keyword} />
+            </span>
+            <span className="search-preview__book__publisher">
+              <EmphasisInString
+                wholeString={book.publisher}
+                emphasis={keyword}
+              />
+            </span>
+          </button>
+        ))}
       </div>
-      <PaginationCircle
-        className="search-preview__pagination"
-        currentPage={page}
-        setCurrentPage={setPage}
-        length={pageCount}
+      <Image
+        className="search-preview__image"
+        width={120}
+        height={180}
+        src={books[selectedIndex]?.image}
+        alt={books[selectedIndex]?.title}
       />
-      <Link
-        className="search-preview__more"
-        to={`/search?search=${encodeURI(keyword)}`}
-      >
-        {`
-           
-          검색 결과 더보기`}
-      </Link>
-    </>
+    </div>
   );
 };
